async req module
