+++ config/redis.yml
@@ -8,6 +8,9 @@ development: &defaults
   rx_store:
     namespace: rx-service
     each_ttl: 1200
+  gi_bill_feedback:
+    namespace: gi_bill_feedback
+    each_ttl: 86400
   sm_store:
     namespace: sm-service
     each_ttl: 1200
